commit FormData()
BACKEND
- Model Post.js => Ajout du champ username dans la bd SQL et dans Post.js sur la fonction createInTable
- Controllers post.js => Ajout de username dans la fonction logique métier createPost
- Routes post.js => Modif des routes endpoints

- Controller user.js => Récupération de userId dans la réponse ligne 80 du backend nécessaire pour le frontend (post et comments nécessitent le userId pour remplir les tables respectives)
- Route user.js => Modif des routes

Multer => Ajout du code + ajout middleware sur les routes nécessaires pour permettre au back de lire données envoyées par le frontend Vue.js

FRONTEND
- axios.post et new FormData => Formatage des données de formulaire transmis via axios, il faut préciser, traduire les données transmises => on a un tableau avec +sieurs types de données différents à envoyer => string, int, images etc....
header 'content-type': 'multipart/form-data' est appliquée pour transfert vers le back des données du form

- index.html + Login + Signup => Ajout CDN fontawesome pour icone user. Modifs couleur label form
- UserAccount: initialisation transfert au back avec axios | FormData()
- PostsComments => Mise en place rendu dynamique + Restructuration de la page + transfert correct des données pour la création de posts avec axios + FormData + Utilisation correct du v-if permettant de montrer le contenu de la page ou pas

+++ P7_GROUPOMANIA/groupo-vue-cli-app/backend/models/Post.js
@@ -2,20 +2,21 @@ const dbmySql = require('../mysqlconnection');  // import configuration de conne
 
 /* FICHIER MODEL POST AVEC TOUTES LES REQUETES SQL CRUD concernant le tableau 'posts' de notre DB mysql
 Requetes sql de l'API pour les 'posts'
-Table post => id | post | userId | date_creation */
+Table post => id | post | userId | username |  date_creation */
 
 // REQUETES CRUD SUR LA TABLE POSTS DE NOTRE DB MYSQL
 
 // Fonction requête sql pour affichage de tous les posts
-exports.getAllInTable = () => {
+exports.getAll = () => {
     return new Promise((resolve, reject) => { 
         // Préparation requête SQL
-        const sql = 'SELECT * FROM posts';
+        // const sql = 'SELECT * FROM posts' changement SQL avec jointure sur table users
+        // faire des tests ici pour enlever le duplicated key id et userId the same
+        const sql = 'SELECT * FROM posts JOIN users ON posts.userId = users.id ORDER BY posts.date_creation DESC';
         // Exécution requête
         dbmySql.query( sql, function(error, results, fields) {
             if (error) reject(error);
             resolve(results);
-            
         });
     })
 };
@@ -23,61 +24,55 @@ exports.getAllInTable = () => {
 
 // Fonction requête sql pour affichage d'un seul post
 // Requete query on selectionne toutes les colonnes correspondant à l'id_post en parametre
-exports.getOneInTable = (id_primary) => { // idPost === primaryId
+exports.getOne = (id_primary) => { // idPost === primaryId
     return new Promise((resolve, reject) => { //gestion asynchrone
-        const sql = 'SELECT * FROM posts WHERE id= ?';
+        // const sql = 'SELECT * FROM posts WHERE id= ?';
+        const sql = 'SELECT * FROM posts JOIN users ON posts.userId = users.id WHERE id=?'
         dbmySql.query( sql , [id_primary] , function(error, results, fields) {
             if (error) reject(error);
             resolve(results);
-            
         })
     })
 };
 
 
 // Fonction requête sql pour CREATION d'une ligne dans la table post
-exports.createInTable = (post, userId) => { // userId est la clé primaire id de user
+// userId est la clé primaire id de user
+exports.create = (post, userId, username) => { 
     return new Promise((resolve, reject) => {
-        //ERREUR 44-17 | 40-12 |
-        const dataInserted = [post, userId]
-        const sql = 'INSERT INTO posts (post, userId) VALUES (?,?)';
+        const dataInserted = [post, userId, username]
+        const sql = 'INSERT INTO posts ( post, userId, username) VALUES (?,?,?)';
         dbmySql.query(sql, dataInserted, function (error, results, fields){
             if (error) reject (error);
             resolve(results);
             console.log(results);
-            
         })
     })
 };
 
 
 // Fonction requête sql pour MODIFIER un post
-exports.updateInTable = ( post, userId, idPost ) => {
+exports.update = ( post, userId, idPost ) => {
     return new Promise((resolve,reject) => {
         //requete sql dans une const
-        const sql = 'UPDATE posts SET post=?, userId=? WHERE id=?';
+        const sql = 'UPDATE posts SET post=? WHERE id=? AND userId=? ';
         let dataUpdated = [post, userId, idPost]
         dbmySql.query( sql, dataUpdated , function (error, results, fields){
             if (error) reject (error);
             resolve(results);
-            
         })
-    
     })
-    
 };
 
 
 // Fonction requête sql pour SUPPRIMER un post
-exports.deleteInTable = (id_post) => { // ds ma db l'id du post est id et non id_post
+exports.delete = (id_post) => { 
     return new Promise((resolve,reject) => {
-        const sql = 'DELETE FROM posts WHERE id= ?';
+        // const sql = 'DELETE FROM posts WHERE id= ?';
+        const sql = 'DELETE FROM posts WHERE id=? AND userId=?';
         dbmySql.query( sql, [id_post], function (error, results, fields){
             if (error) reject (error);
             resolve(results);
-            
         })
     })
 };
-
-
